Fix some bugs from valgrind

+++ hdkey/HDkey.c
@@ -471,7 +471,7 @@ static int getv_sub_privatekey_publickey(HDKey *hdkey, HDKey *derivedkey, int de
     memcpy(derivedkey->prvChainCode, chaincode.u8, sizeof(chaincode.u8));
     memcpy(derivedkey->pubChainCode, chaincode.u8, sizeof(chaincode.u8));
 
-    derivedkey->depth += depth;
+    derivedkey->depth += (uint8_t)depth;
 
     var_clean(&chaincode);
     var_clean(&secret);
